envelope count check  & tests for envelope_checker
